requirements.txt
fastapi>=0.112.1
sqlalchemy>=2.0.32
scikit-learn>=1.5.1
numpy>=2.1.0
pandas>=2.2.2
uvicorn>=0.30.6
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
mlflow==2.15.1
boto3==1.35.1
matplotlib==3.9.0
seaborn>=0.13.2